+++ Assets/Scripts/GameController/VersusModeManager.cs
@@ -63,7 +63,7 @@ public class VersusModeManager : GameModeManager
 
         getWinner();
 
-        controller.data.points += (score[0] + score[1] + score[2] + score[3]);
+        controller.data.points += (int)(score[0] + score[1] + score[2] + score[3]);
         FileManager.saveData(controller.data);
     }
